fix: change Size field type from int32 to int64
使用openapi-generator重新生成api client,
新版本将integer类型生成为int64而不是int32

int32会导致数值溢出

+++ pkg/types/types.go
@@ -20,7 +20,7 @@ type LayerInfo struct {
 	Module        string                `json:"module"`
 	Name          string                `json:"name"`
 	Runtime       string                `json:"runtime"`
-	Size          int                   `json:"size"`
+	Size          int64                 `json:"size"`
 	Version       string                `json:"version"`
 	Channel       string                `json:"channel"`
 	SchemaVersion string                `json:"schema_version"`
